+++ poolMock/PooledObject.java
@@ -1,0 +1,4 @@
+package com.ibm.mq.spring.boot;
+
+class PooledObject {
+}